Retry when former leader failed.
If the former leader is no longer the leader, but we think that they are, when we attempt to query it with the session object, we will receive an error message.  That doesn't mean that we can no longer proceed, it only means we have to find the new leader by attempting one of the other nodes.